Functional Note w/ Logic
The .ino code plays the full note with repeated middle section to show how it must be done. The flute A4 note that was referenced in the code has been updated to reduce static noises.
There are still clicking sounds between swapping audio sources (beg -> mid, mid -> mid, mid -> end), but that's something that must be addressed in the audio library eventually, not by editing the actual wav files.
This commit included the audio library will all the edits required to support the current state.